experimental/requirements.txt
flask>=3.0.0
flask-cors>=4.0.1
gnews>=0.3.7
google-api-python-client>=2.125.0
google-auth>=2.29.0
gtts>=2.5.1
nltk>=3.8.1
playsound==1.2.2
python-dotenv>=1.0.1
pytz>=2024.1
requests>=2.32.0
simpleaudio>=1.0.4
SpeechRecognition>=3.10.4
transformers>=4.40.0
